Move all util methods and types into own files.
