+++ reddwarf/utils/pca.py
@@ -1,0 +1,66 @@
+import pandas as pd
+import numpy as np
+from sklearn.decomposition import PCA
+from reddwarf.utils.matrix import VoteMatrix, impute_missing_votes
+from typing import Tuple
+
+
+def run_pca(
+        vote_matrix: VoteMatrix,
+        n_components: int = 2,
+) -> Tuple[ pd.DataFrame, np.ndarray, np.ndarray ]:
+    """
+    Process a prepared vote matrix to be imputed and return projected participant data,
+    as well as eigenvectors and eigenvalues.
+
+    The vote matrix should not yet be imputed, as this will happen within the method.
+
+    Args:
+        vote_matrix (pd.DataFrame): A vote matrix of data. Non-imputed values are expected.
+        n_components (int): Number n of principal components to decompose the `vote_matrix` into.
+
+    Returns:
+        projected_data (pd.DataFrame): A dataframe of projected xy coordinates for each `vote_matrix` row.
+        eigenvectors (List[List[float]]): Principal `n` components, one per row.
+        eigenvalues (List[float]): Explained variance,  one per row.
+    """
+    imputed_matrix = impute_missing_votes(vote_matrix)
+
+    pca = PCA(n_components=n_components) ## pca is apparently different, it wants
+    pca.fit(imputed_matrix) ## .T transposes the matrix (flips it)
+
+    eigenvectors = pca.components_
+    eigenvalues = pca.explained_variance_
+
+    # Project participant vote data onto 2D using eigenvectors.
+    projected_data = pca.transform(imputed_matrix)
+    projected_data = pd.DataFrame(projected_data, index=imputed_matrix.index, columns=["x", "y"])
+    projected_data.index.name = "participant_id"
+
+    return projected_data, eigenvectors, eigenvalues
+
+def scale_projected_data(
+        projected_data: pd.DataFrame,
+        vote_matrix: VoteMatrix
+) -> pd.DataFrame:
+    """
+    Scale projected participant xy points based on vote matrix, to account for any small number of
+    votes by a participant and prevent those participants from bunching up in the center.
+
+    Args:
+        projected_data (pd.DataFrame): the project xy coords of participants.
+        vote_matrix (VoteMatrix): the processed vote matrix data frame, from which to generate scaling factors.
+
+    Returns:
+        scaled_projected_data (pd.DataFrame): The coord data rescaled based on participant votes.
+    """
+    total_active_comment_count = vote_matrix.shape[1]
+    participant_vote_counts = vote_matrix.count(axis="columns")
+    # Ref: https://hyp.is/x6nhItMMEe-v1KtYFgpOiA/gwern.net/doc/sociology/2021-small.pdf
+    # Ref: https://github.com/compdemocracy/polis/blob/15aa65c9ca9e37ecf57e2786d7d81a4bd4ad37ef/math/src/polismath/math/pca.clj#L155-L156
+    participant_scaling_coeffs = np.sqrt(total_active_comment_count / participant_vote_counts).values
+    # See: https://numpy.org/doc/stable/reference/generated/numpy.reshape.html
+    # Reshape scaling_coeffs list to match the shape of projected_data matrix
+    participant_scaling_coeffs = np.reshape(participant_scaling_coeffs, (-1, 1))
+
+    return projected_data * participant_scaling_coeffs
